@@ -97,7 +97,7 @@ void Node::loadFromByteArray(const byte* ptr)
 		}
 		else
 		{
-			m_pData[u32Child] = 0;
+			m_pData[u32Child] = nullptr;
 		}
 
 		//m_nodeMBR.combineRegion(*(m_ptrMBR[u32Child]));
@@ -196,10 +196,10 @@ void Node::getChildShape(uint32_t index, IShape** out) const
 void Node::getChildData(uint32_t index, uint32_t& length, byte** data) const
 {
 	if (index >= m_children) throw Tools::IndexOutOfBoundsException(index);
-	if (m_pData[index] == NULL)
+	if (m_pData[index] == nullptr)
 	{
 		length = 0;
-		data = NULL;
+		data = nullptr;
 	}
 	else
 	{
@@ -227,19 +227,8 @@ bool Node::isIndex() const
 // Internal
 //
 
-Node::Node() :
-	m_pTree(0),
-	m_level(0),
-	m_identifier(-1),
-	m_children(0),
-	m_capacity(0),
-	m_pData(0),
-	m_ptrMBR(0),
-	m_pIdentifier(0),
-	m_pDataLength(0),
-	m_totalDataLength(0)
-{
-}
+Node::Node()
+= default;
 
 Node::Node(SpatialIndex::RTree::RTree* pTree, id_type id, uint32_t level, uint32_t capacity) :
 	m_pTree(pTree),
@@ -247,10 +236,10 @@ Node::Node(SpatialIndex::RTree::RTree* pTree, id_type id, uint32_t level, uint32
 	m_identifier(id),
 	m_children(0),
 	m_capacity(capacity),
-	m_pData(0),
-	m_ptrMBR(0),
-	m_pIdentifier(0),
-	m_pDataLength(0),
+	m_pData(nullptr),
+	m_ptrMBR(nullptr),
+	m_pIdentifier(nullptr),
+	m_pDataLength(nullptr),
 	m_totalDataLength(0)
 {
 	m_nodeMBR.makeInfinite(m_pTree->m_dimension);
@@ -274,11 +263,11 @@ Node::Node(SpatialIndex::RTree::RTree* pTree, id_type id, uint32_t level, uint32
 
 Node::~Node()
 {
-	if (m_pData != 0)
+	if (m_pData != nullptr)
 	{
 		for (uint32_t u32Child = 0; u32Child < m_children; ++u32Child)
 		{
-			if (m_pData[u32Child] != 0) delete[] m_pData[u32Child];
+			if (m_pData[u32Child] != nullptr) delete[] m_pData[u32Child];
 		}
 
 		delete[] m_pData;
@@ -318,7 +307,7 @@ void Node::deleteEntry(uint32_t index)
 	RegionPtr ptrR = m_ptrMBR[index];
 
 	m_totalDataLength -= m_pDataLength[index];
-	if (m_pData[index] != 0) delete[] m_pData[index];
+	if (m_pData[index] != nullptr) delete[] m_pData[index];
 
 	if (m_children > 1 && index != m_children - 1)
 	{
@@ -385,14 +374,14 @@ bool Node::insertData(uint32_t dataLength, byte* pData, Region& mbr, id_type id,
 		uint32_t lReinsert = static_cast<uint32_t>(vReinsert.size());
 		uint32_t lKeep = static_cast<uint32_t>(vKeep.size());
 
-		byte** reinsertdata = 0;
-		RegionPtr* reinsertmbr = 0;
-		id_type* reinsertid = 0;
-		uint32_t* reinsertlen = 0;
-		byte** keepdata = 0;
-		RegionPtr* keepmbr = 0;
-		id_type* keepid = 0;
-		uint32_t* keeplen = 0;
+		byte** reinsertdata = nullptr;
+		RegionPtr* reinsertmbr = nullptr;
+		id_type* reinsertid = nullptr;
+		uint32_t* reinsertlen = nullptr;
+		byte** keepdata = nullptr;
+		RegionPtr* keepmbr = nullptr;
+		id_type* keepid = nullptr;
+		uint32_t* keeplen = nullptr;
 
 		try
 		{
@@ -471,7 +460,7 @@ bool Node::insertData(uint32_t dataLength, byte* pData, Region& mbr, id_type id,
 		id_type cParent = pathBuffer.top(); pathBuffer.pop();
 		NodePtr ptrN = m_pTree->readNode(cParent);
 		Index* p = static_cast<Index*>(ptrN.get());
-		p->adjustTree(this, pathBuffer);
+		p->adjustTree(this, pathBuffer, true);
 
 		for (cIndex = 0; cIndex < lReinsert; ++cIndex)
 		{
@@ -504,7 +493,7 @@ bool Node::insertData(uint32_t dataLength, byte* pData, Region& mbr, id_type id,
 			m_pTree->writeNode(nn.get());
 
 			NodePtr ptrR = m_pTree->m_indexPool.acquire();
-			if (ptrR.get() == 0)
+			if (ptrR.get() == nullptr)
 			{
 				ptrR = NodePtr(new Index(m_pTree, m_pTree->m_rootID, m_level + 1), &(m_pTree->m_indexPool));
 			}
@@ -516,8 +505,8 @@ bool Node::insertData(uint32_t dataLength, byte* pData, Region& mbr, id_type id,
 				ptrR->m_nodeMBR = m_pTree->m_infiniteRegion;
 			}
 
-			ptrR->insertEntry(0, 0, n->m_nodeMBR, n->m_identifier);
-			ptrR->insertEntry(0, 0, nn->m_nodeMBR, nn->m_identifier);
+			ptrR->insertEntry(0, nullptr, n->m_nodeMBR, n->m_identifier);
+			ptrR->insertEntry(0, nullptr, nn->m_nodeMBR, nn->m_identifier);
 
 			m_pTree->writeNode(ptrR.get());
 
@@ -589,15 +578,21 @@ void Node::reinsertData(uint32_t dataLength, byte* pData, Region& mbr, id_type i
 
 	uint32_t cCount;
 
-	for (cCount = 0; cCount < cReinsert; ++cCount)
+	for (cCount = 0; cCount < m_capacity + 1; ++cCount)
 	{
-		reinsert.push_back(v[cCount]->m_index);
-		delete v[cCount];
-	}
-
-	for (cCount = cReinsert; cCount < m_capacity + 1; ++cCount)
-	{
-		keep.push_back(v[cCount]->m_index);
+		if (cCount < m_capacity + 1 - cReinsert)
+		{
+			// Keep all but cReinsert nodes
+			keep.push_back(v[cCount]->m_index);
+		}
+		else
+		{
+			// Remove cReinsert nodes which will be
+			// reinserted into the tree. Since our array
+			// is already sorted in ascending order this
+			// matches the order suggested in the paper.
+			reinsert.push_back(v[cCount]->m_index);
+		}
 		delete v[cCount];
 	}
 
@@ -760,8 +755,8 @@ void Node::rtreeSplit(uint32_t dataLength, byte* pData, Region& mbr, id_type id,
 
 void Node::rstarSplit(uint32_t dataLength, byte* pData, Region& mbr, id_type id, std::vector<uint32_t>& group1, std::vector<uint32_t>& group2)
 {
-	RstarSplitEntry** dataLow = 0;
-	RstarSplitEntry** dataHigh = 0;
+	RstarSplitEntry** dataLow = nullptr;
+	RstarSplitEntry** dataHigh = nullptr;
 
 	try
 	{
@@ -1027,6 +1022,27 @@ void Node::condenseTree(std::stack<NodePtr>& toReinsert, std::stack<id_type>& pa
 			m_pTree->m_stats.m_u32TreeHeight -= 1;
 			// HACK: pending deleteNode for deleted child will decrease nodesInLevel, later on.
 			m_pTree->m_stats.m_nodesInLevel[m_pTree->m_stats.m_u32TreeHeight - 1] = 2;
+		}
+		else
+		{
+			// due to data removal.
+			if (m_pTree->m_bTightMBRs)
+			{
+				for (uint32_t cDim = 0; cDim < m_nodeMBR.m_dimension; ++cDim)
+				{
+					m_nodeMBR.m_pLow[cDim] = std::numeric_limits<double>::max();
+					m_nodeMBR.m_pHigh[cDim] = -std::numeric_limits<double>::max();
+
+					for (uint32_t u32Child = 0; u32Child < m_children; ++u32Child)
+					{
+						m_nodeMBR.m_pLow[cDim] = std::min(m_nodeMBR.m_pLow[cDim], m_ptrMBR[u32Child]->m_pLow[cDim]);
+						m_nodeMBR.m_pHigh[cDim] = std::max(m_nodeMBR.m_pHigh[cDim], m_ptrMBR[u32Child]->m_pHigh[cDim]);
+					}
+				}
+			}
+
+            // write parent node back to storage.
+			m_pTree->writeNode(this);
 		}
 	}
 	else
