@@ -5,7 +5,7 @@
  * Copyright (c) 2002, Marios Hadjieleftheriou
  *
  * All rights reserved.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation
@@ -30,6 +30,10 @@
 
 // For checking if a file exists - hobu
 #include <sys/stat.h>
+
+#ifdef WIN32
+#define stat _stat64
+#endif
 
 #include <spatialindex/SpatialIndex.h>
 #include "DiskStorageManager.h"
@@ -121,7 +125,7 @@ SpatialIndex::IStorageManager* SpatialIndex::StorageManager::loadDiskStorageMana
 	return returnDiskStorageManager(ps);
 }
 
-DiskStorageManager::DiskStorageManager(Tools::PropertySet& ps) : m_pageSize(0), m_nextPage(-1), m_buffer(0)
+DiskStorageManager::DiskStorageManager(Tools::PropertySet& ps) : m_pageSize(0), m_nextPage(-1), m_buffer(nullptr)
 {
 	Tools::Variant var;
 
@@ -129,7 +133,7 @@ DiskStorageManager::DiskStorageManager(Tools::PropertySet& ps) : m_pageSize(0), 
 	bool bOverwrite = false;
 	bool bFileExists = false;
 	std::streamoff length = 0;
-	
+
 	var = ps.getProperty("Overwrite");
 
 	if (var.m_varType != Tools::VT_EMPTY)
@@ -171,7 +175,7 @@ DiskStorageManager::DiskStorageManager(Tools::PropertySet& ps) : m_pageSize(0), 
 			m_dataFile.open(sDataFile.c_str(), mode);
 
 			if (m_indexFile.fail() || m_dataFile.fail())
-				throw Tools::IllegalArgumentException("SpatialIndex::DiskStorageManager: Index/Data file cannot be read/writen.");
+				throw Tools::IllegalArgumentException("SpatialIndex::DiskStorageManager: Index/Data file cannot be read/written.");
 		}
 		else
 		{
@@ -283,10 +287,10 @@ DiskStorageManager::~DiskStorageManager()
 	flush();
 	m_indexFile.close();
 	m_dataFile.close();
-	if (m_buffer != 0) delete[] m_buffer;
+	if (m_buffer != nullptr) delete[] m_buffer;
 
-	std::map<id_type, Entry*>::iterator it;
-	for (it = m_pageIndex.begin(); it != m_pageIndex.end(); ++it) delete (*it).second;
+	for (auto& v: m_pageIndex)
+		delete v.second;
 }
 
 void DiskStorageManager::flush()
