@@ -112,7 +112,7 @@ void Node::loadFromByteArray(const byte* ptr)
 		}
 		else
 		{
-			m_pData[cChild] = 0;
+			m_pData[cChild] = nullptr;
 		}
 
 		//m_nodeMBR.combineRegion(*(m_ptrMBR[cChild]));
@@ -229,10 +229,10 @@ void Node::getChildShape(uint32_t index, IShape** out) const
 void Node::getChildData(uint32_t index, uint32_t& length, byte** data) const
 {
 	if (index >= m_children) throw Tools::IndexOutOfBoundsException(index);
-	if (m_pData[index] == NULL)
+	if (m_pData[index] == nullptr)
 	{
 		length = 0;
-		data = NULL;
+		data = nullptr;
 	}
 	else
 	{
@@ -260,19 +260,8 @@ bool Node::isIndex() const
 // Internal
 //
 
-Node::Node() :
-	m_pTree(0),
-	m_level(0),
-	m_identifier(-1),
-	m_children(0),
-	m_capacity(0),
-	m_pData(0),
-	m_ptrMBR(0),
-	m_pIdentifier(0),
-	m_pDataLength(0),
-	m_totalDataLength(0)
-{
-}
+Node::Node()
+= default;
 
 Node::Node(SpatialIndex::TPRTree::TPRTree* pTree, id_type id, uint32_t level, uint32_t capacity) :
 	m_pTree(pTree),
@@ -280,10 +269,10 @@ Node::Node(SpatialIndex::TPRTree::TPRTree* pTree, id_type id, uint32_t level, ui
 	m_identifier(id),
 	m_children(0),
 	m_capacity(capacity),
-	m_pData(0),
-	m_ptrMBR(0),
-	m_pIdentifier(0),
-	m_pDataLength(0),
+	m_pData(nullptr),
+	m_ptrMBR(nullptr),
+	m_pIdentifier(nullptr),
+	m_pDataLength(nullptr),
 	m_totalDataLength(0)
 {
 	m_nodeMBR.makeInfinite(m_pTree->m_dimension);
@@ -307,11 +296,11 @@ Node::Node(SpatialIndex::TPRTree::TPRTree* pTree, id_type id, uint32_t level, ui
 
 Node::~Node()
 {
-	if (m_pData != 0)
+	if (m_pData != nullptr)
 	{
 		for (uint32_t cChild = 0; cChild < m_children; ++cChild)
 		{
-			if (m_pData[cChild] != 0) delete[] m_pData[cChild];
+			if (m_pData[cChild] != nullptr) delete[] m_pData[cChild];
 		}
 
 		delete[] m_pData;
@@ -405,7 +394,7 @@ void Node::deleteEntry(uint32_t index)
 	MovingRegionPtr ptrR = m_ptrMBR[index];
 
 	m_totalDataLength -= m_pDataLength[index];
-	if (m_pData[index] != 0) delete[] m_pData[index];
+	if (m_pData[index] != nullptr) delete[] m_pData[index];
 
 	if (m_children > 1 && index != m_children - 1)
 	{
@@ -481,14 +470,14 @@ bool Node::insertData(uint32_t dataLength, byte* pData, MovingRegion& mbr, id_ty
 		uint32_t lReinsert = static_cast<uint32_t>(vReinsert.size());
 		uint32_t lKeep = static_cast<uint32_t>(vKeep.size());
 
-		byte** reinsertdata = 0;
-		MovingRegionPtr* reinsertmbr = 0;
-		id_type* reinsertid = 0;
-		uint32_t* reinsertlen = 0;
-		byte** keepdata = 0;
-		MovingRegionPtr* keepmbr = 0;
-		id_type* keepid = 0;
-		uint32_t* keeplen = 0;
+		byte** reinsertdata = nullptr;
+		MovingRegionPtr* reinsertmbr = nullptr;
+		id_type* reinsertid = nullptr;
+		uint32_t* reinsertlen = nullptr;
+		byte** keepdata = nullptr;
+		MovingRegionPtr* keepmbr = nullptr;
+		id_type* keepid = nullptr;
+		uint32_t* keeplen = nullptr;
 
 		try
 		{
@@ -627,7 +616,7 @@ bool Node::insertData(uint32_t dataLength, byte* pData, MovingRegion& mbr, id_ty
 			m_pTree->writeNode(nn.get());
 
 			NodePtr ptrR = m_pTree->m_indexPool.acquire();
-			if (ptrR.get() == 0)
+			if (ptrR.get() == nullptr)
 			{
 				ptrR = NodePtr(new Index(m_pTree, m_pTree->m_rootID, m_level + 1), &(m_pTree->m_indexPool));
 			}
@@ -639,8 +628,8 @@ bool Node::insertData(uint32_t dataLength, byte* pData, MovingRegion& mbr, id_ty
 				ptrR->m_nodeMBR = m_pTree->m_infiniteRegion;
 			}
 
-			ptrR->insertEntry(0, 0, n->m_nodeMBR, n->m_identifier);
-			ptrR->insertEntry(0, 0, nn->m_nodeMBR, nn->m_identifier);
+			ptrR->insertEntry(0, nullptr, n->m_nodeMBR, n->m_identifier);
+			ptrR->insertEntry(0, nullptr, nn->m_nodeMBR, nn->m_identifier);
 
 			m_pTree->writeNode(ptrR.get());
 
@@ -886,10 +875,10 @@ void Node::rtreeSplit(uint32_t dataLength, byte* pData, Region& mbr, id_type id,
 
 void Node::rstarSplit(uint32_t dataLength, byte* pData, MovingRegion& mbr, id_type id, std::vector<uint32_t>& group1, std::vector<uint32_t>& group2)
 {
-	RstarSplitEntry** dataLow = 0;
-	RstarSplitEntry** dataHigh = 0;
-	RstarSplitEntry** dataVLow = 0;
-	RstarSplitEntry** dataVHigh = 0;
+	RstarSplitEntry** dataLow = nullptr;
+	RstarSplitEntry** dataHigh = nullptr;
+	RstarSplitEntry** dataVLow = nullptr;
+	RstarSplitEntry** dataVHigh = nullptr;
 
 	try
 	{
