@@ -29,8 +29,8 @@
 
 #ifndef SIDX_VERSION_MAJOR
 #define SIDX_VERSION_MAJOR    1
-#define SIDX_VERSION_MINOR    8
-#define SIDX_VERSION_REV      5
+#define SIDX_VERSION_MINOR    9
+#define SIDX_VERSION_REV      0
 #define SIDX_VERSION_BUILD    0
 #endif
 
@@ -39,10 +39,10 @@
 #endif
 
 #ifndef SIDX_RELEASE_DATE
-#define SIDX_RELEASE_DATE     20141101
+#define SIDX_RELEASE_DATE     20170302
 #endif
 
 #ifndef SIDX_RELEASE_NAME
-#define SIDX_RELEASE_NAME     "1.8.5"
+#define SIDX_RELEASE_NAME     "1.9.0"
 #endif
 
