@@ -35,8 +35,7 @@ using namespace SpatialIndex;
 using namespace SpatialIndex::MVRTree;
 
 Leaf::~Leaf()
-{
-}
+= default;
 
 Leaf::Leaf(SpatialIndex::MVRTree::MVRTree* pTree, id_type id): Node(pTree, id, 0, pTree->m_leafCapacity)
 {
@@ -86,8 +85,8 @@ void Leaf::split(
 	pLeft = m_pTree->m_leafPool.acquire();
 	pRight = m_pTree->m_leafPool.acquire();
 
-	if (pLeft.get() == 0) pLeft = NodePtr(new Leaf(m_pTree, -1), &(m_pTree->m_leafPool));
-	if (pRight.get() == 0) pRight = NodePtr(new Leaf(m_pTree, -1), &(m_pTree->m_leafPool));
+	if (pLeft.get() == nullptr) pLeft = NodePtr(new Leaf(m_pTree, -1), &(m_pTree->m_leafPool));
+	if (pRight.get() == nullptr) pRight = NodePtr(new Leaf(m_pTree, -1), &(m_pTree->m_leafPool));
 
 	pLeft->m_nodeMBR = m_pTree->m_infiniteRegion;
 	pRight->m_nodeMBR = m_pTree->m_infiniteRegion;
@@ -98,13 +97,13 @@ void Leaf::split(
 	{
 		pLeft->insertEntry(m_pDataLength[g1[cIndex]], m_pData[g1[cIndex]], *(m_ptrMBR[g1[cIndex]]), m_pIdentifier[g1[cIndex]]);
 		// we don't want to delete the data array from this node's destructor!
-		m_pData[g1[cIndex]] = 0;
+		m_pData[g1[cIndex]] = nullptr;
 	}
 
 	for (cIndex = 0; cIndex < g2.size(); ++cIndex)
 	{
 		pRight->insertEntry(m_pDataLength[g2[cIndex]], m_pData[g2[cIndex]], *(m_ptrMBR[g2[cIndex]]), m_pIdentifier[g2[cIndex]]);
 		// we don't want to delete the data array from this node's destructor!
-		m_pData[g2[cIndex]] = 0;
+		m_pData[g2[cIndex]] = nullptr;
 	}
 }
