@@ -5,7 +5,7 @@
  * Copyright (c) 2002, Marios Hadjieleftheriou
  *
  * All rights reserved.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
  * to deal in the Software without restriction, including without limitation
@@ -39,7 +39,7 @@
 using namespace SpatialIndex::MVRTree;
 
 SpatialIndex::MVRTree::Data::Data(uint32_t len, byte* pData, TimeRegion& r, id_type id)
-	: m_id(id), m_region(r), m_pData(0), m_dataLength(len)
+	: m_id(id), m_region(r), m_pData(nullptr), m_dataLength(len)
 {
 	if (m_dataLength > 0)
 	{
@@ -71,7 +71,7 @@ void SpatialIndex::MVRTree::Data::getShape(IShape** out) const
 void SpatialIndex::MVRTree::Data::getData(uint32_t& len, byte** data) const
 {
 	len = m_dataLength;
-	*data = 0;
+	*data = nullptr;
 
 	if (m_dataLength > 0)
 	{
@@ -95,7 +95,7 @@ void SpatialIndex::MVRTree::Data::loadFromByteArray(const byte* ptr)
 	ptr += sizeof(id_type);
 
 	delete[] m_pData;
-	m_pData = 0;
+	m_pData = nullptr;
 
 	memcpy(&m_dataLength, ptr, sizeof(uint32_t));
 	ptr += sizeof(uint32_t);
@@ -114,7 +114,7 @@ void SpatialIndex::MVRTree::Data::storeToByteArray(byte** data, uint32_t& len)
 {
 	// it is thread safe this way.
 	uint32_t regionsize;
-	byte* regiondata = 0;
+	byte* regiondata = nullptr;
 	m_region.storeToByteArray(&regiondata, regionsize);
 
 	len = sizeof(id_type) + sizeof(uint32_t) + m_dataLength + regionsize;
@@ -219,9 +219,6 @@ SpatialIndex::MVRTree::MVRTree::MVRTree(IStorageManager& sm, Tools::PropertySet&
 	m_indexPool(100),
 	m_leafPool(100)
 {
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
 
 	Tools::Variant var = ps.getProperty("IndexIdentifier");
 	if (var.m_varType != Tools::VT_EMPTY)
@@ -244,10 +241,6 @@ SpatialIndex::MVRTree::MVRTree::MVRTree(IStorageManager& sm, Tools::PropertySet&
 
 SpatialIndex::MVRTree::MVRTree::~MVRTree()
 {
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	storeHeader();
 }
 
@@ -259,12 +252,8 @@ void SpatialIndex::MVRTree::MVRTree::insertData(uint32_t len, const byte* pData,
 {
 	if (shape.getDimension() != m_dimension) throw Tools::IllegalArgumentException("insertData: Shape has the wrong number of dimensions.");
 	const Tools::IInterval* ti = dynamic_cast<const Tools::IInterval*>(&shape);
-	if (ti == 0) throw Tools::IllegalArgumentException("insertData: Shape does not support the Tools::IInterval interface.");
+	if (ti == nullptr) throw Tools::IllegalArgumentException("insertData: Shape does not support the Tools::IInterval interface.");
 	if (ti->getLowerBound() < m_currentTime) throw Tools::IllegalArgumentException("insertData: Shape start time is older than tree current time.");
-
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
 
 	// convert the shape into a TimeRegion (R-Trees index regions only; i.e., approximations of the shapes).
 	Region mbrold;
@@ -278,7 +267,7 @@ void SpatialIndex::MVRTree::MVRTree::insertData(uint32_t len, const byte* pData,
 	mbr->m_startTime = ti->getLowerBound();
 	mbr->m_endTime = std::numeric_limits<double>::max();
 
-	byte* buffer = 0;
+	byte* buffer = nullptr;
 
 	if (len > 0)
 	{
@@ -294,11 +283,7 @@ bool SpatialIndex::MVRTree::MVRTree::deleteData(const IShape& shape, id_type id)
 {
 	if (shape.getDimension() != m_dimension) throw Tools::IllegalArgumentException("deleteData: Shape has the wrong number of dimensions.");
 	const Tools::IInterval* ti = dynamic_cast<const Tools::IInterval*>(&shape);
-	if (ti == 0) throw Tools::IllegalArgumentException("deleteData: Shape does not support the Tools::IInterval interface.");
-
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
+	if (ti == nullptr) throw Tools::IllegalArgumentException("deleteData: Shape does not support the Tools::IInterval interface.");
 
 	Region mbrold;
 	shape.getMBR(mbrold);
@@ -314,6 +299,12 @@ bool SpatialIndex::MVRTree::MVRTree::deleteData(const IShape& shape, id_type id)
 	bool ret = deleteData_impl(*mbr, id);
 
 	return ret;
+}
+
+
+void SpatialIndex::MVRTree::MVRTree::internalNodesQuery(const IShape& query, IVisitor& v)
+{
+	throw Tools::IllegalStateException("internalNodesQuery: not impelmented yet.");
 }
 
 void SpatialIndex::MVRTree::MVRTree::containsWhatQuery(const IShape& query, IVisitor& v)
@@ -332,7 +323,7 @@ void SpatialIndex::MVRTree::MVRTree::pointLocationQuery(const Point& query, IVis
 {
 	if (query.m_dimension != m_dimension) throw Tools::IllegalArgumentException("pointLocationQuery: Shape has the wrong number of dimensions.");
 	const Tools::IInterval* ti = dynamic_cast<const Tools::IInterval*>(&query);
-	if (ti == 0) throw Tools::IllegalArgumentException("pointLocationQuery: Shape does not support the Tools::IInterval interface.");
+	if (ti == nullptr) throw Tools::IllegalArgumentException("pointLocationQuery: Shape does not support the Tools::IInterval interface.");
 	TimeRegion r(query, query, *ti);
 	rangeQuery(IntersectionQuery, r, v);
 }
@@ -356,10 +347,6 @@ void SpatialIndex::MVRTree::MVRTree::selfJoinQuery(const IShape&, IVisitor&)
 
 void SpatialIndex::MVRTree::MVRTree::queryStrategy(IQueryStrategy& qs)
 {
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	id_type next = m_roots[m_roots.size() - 1].m_id;
 	bool hasNext = true;
 
@@ -453,6 +440,10 @@ void SpatialIndex::MVRTree::MVRTree::getIndexProperties(Tools::PropertySet& out)
 	var.m_varType = Tools::VT_DOUBLE;
 	var.m_val.dblVal = m_versionUnderflow;
 	out.setProperty("VersionUnderflow", var);
+
+	var.m_varType = Tools::VT_LONGLONG;
+	var.m_val.llVal = m_headerID;
+	out.setProperty("IndexIdentifier", var);
 }
 
 void SpatialIndex::MVRTree::MVRTree::addCommand(ICommand* pCommand, CommandType ct)
@@ -460,13 +451,13 @@ void SpatialIndex::MVRTree::MVRTree::addCommand(ICommand* pCommand, CommandType 
 	switch (ct)
 	{
 		case CT_NODEREAD:
-			m_readNodeCommands.push_back(Tools::SmartPointer<ICommand>(pCommand));
+			m_readNodeCommands.push_back(std::shared_ptr<ICommand>(pCommand));
 			break;
 		case CT_NODEWRITE:
-			m_writeNodeCommands.push_back(Tools::SmartPointer<ICommand>(pCommand));
+			m_writeNodeCommands.push_back(std::shared_ptr<ICommand>(pCommand));
 			break;
 		case CT_NODEDELETE:
-			m_deleteNodeCommands.push_back(Tools::SmartPointer<ICommand>(pCommand));
+			m_deleteNodeCommands.push_back(std::shared_ptr<ICommand>(pCommand));
 			break;
 	}
 }
@@ -564,6 +555,11 @@ bool SpatialIndex::MVRTree::MVRTree::isIndexValid()
 void SpatialIndex::MVRTree::MVRTree::getStatistics(IStatistics** out) const
 {
 	*out = new Statistics(m_stats);
+}
+
+void SpatialIndex::MVRTree::MVRTree::flush()
+{
+	storeHeader();
 }
 
 void SpatialIndex::MVRTree::MVRTree::initNew(Tools::PropertySet& ps)
@@ -741,7 +737,7 @@ void SpatialIndex::MVRTree::MVRTree::initNew(Tools::PropertySet& ps)
 	root.m_nodeMBR.m_startTime = 0.0;
 	root.m_nodeMBR.m_endTime = std::numeric_limits<double>::max();
 	writeNode(&root);
-	m_roots.push_back(RootEntry(root.m_identifier, root.m_nodeMBR.m_startTime, root.m_nodeMBR.m_endTime));
+	m_roots.emplace_back(root.m_identifier, root.m_nodeMBR.m_startTime, root.m_nodeMBR.m_endTime);
 
 	storeHeader();
 }
@@ -961,7 +957,7 @@ void SpatialIndex::MVRTree::MVRTree::storeHeader()
 void SpatialIndex::MVRTree::MVRTree::loadHeader()
 {
 	uint32_t headerSize;
-	byte* header = 0;
+	byte* header = nullptr;
 	m_pStorageManager->loadByteArray(m_headerID, headerSize, &header);
 
 	byte* ptr = header;
@@ -1106,7 +1102,7 @@ bool SpatialIndex::MVRTree::MVRTree::deleteData_impl(const TimeRegion& mbr, id_t
 		root.relinquish();
 	}
 
-	if (l.get() != 0)
+	if (l.get() != nullptr)
 	{
 		l->deleteData(id, mbr.m_endTime, pathBuffer);
 		--(m_stats.m_u64Data);
@@ -1182,7 +1178,7 @@ SpatialIndex::MVRTree::NodePtr SpatialIndex::MVRTree::MVRTree::readNode(id_type 
 		else if (nodeType == PersistentLeaf) n = m_leafPool.acquire();
 		else throw Tools::IllegalStateException("readNode: failed reading the correct node type information");
 
-		if (n.get() == 0)
+		if (n.get() == nullptr)
 		{
 			if (nodeType == PersistentIndex) n = NodePtr(new Index(this, -1, 0), &m_indexPool);
 			else if (nodeType == PersistentLeaf) n = NodePtr(new Leaf(this, -1), &m_leafPool);
@@ -1236,11 +1232,7 @@ void SpatialIndex::MVRTree::MVRTree::rangeQuery(RangeQueryType type, const IShap
 	// FIXME: I am not using ITimeShape yet, even though I should.
 
 	const Tools::IInterval* ti = dynamic_cast<const Tools::IInterval*>(&query);
-	if (ti == 0) throw Tools::IllegalArgumentException("rangeQuery: Shape does not support the Tools::IInterval interface.");
-
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
+	if (ti == nullptr) throw Tools::IllegalArgumentException("rangeQuery: Shape does not support the Tools::IInterval interface.");
 
 	std::set<id_type> visitedNodes;
 	std::set<id_type> visitedData;
