@@ -30,6 +30,7 @@
 #include <limits>
 
 #include <spatialindex/SpatialIndex.h>
+#include <spatialindex/capi/IdVisitor.h>
 #include "Node.h"
 #include "Leaf.h"
 #include "Index.h"
@@ -40,7 +41,7 @@ using namespace SpatialIndex::RTree;
 using namespace SpatialIndex;
 
 SpatialIndex::RTree::Data::Data(uint32_t len, byte* pData, Region& r, id_type id)
-	: m_id(id), m_region(r), m_pData(0), m_dataLength(len)
+	: m_id(id), m_region(r), m_pData(nullptr), m_dataLength(len)
 {
 	if (m_dataLength > 0)
 	{
@@ -72,7 +73,7 @@ void SpatialIndex::RTree::Data::getShape(IShape** out) const
 void SpatialIndex::RTree::Data::getData(uint32_t& len, byte** data) const
 {
 	len = m_dataLength;
-	*data = 0;
+	*data = nullptr;
 
 	if (m_dataLength > 0)
 	{
@@ -96,7 +97,7 @@ void SpatialIndex::RTree::Data::loadFromByteArray(const byte* ptr)
 	ptr += sizeof(id_type);
 
 	delete[] m_pData;
-	m_pData = 0;
+	m_pData = nullptr;
 
 	memcpy(&m_dataLength, ptr, sizeof(uint32_t));
 	ptr += sizeof(uint32_t);
@@ -115,7 +116,7 @@ void SpatialIndex::RTree::Data::storeToByteArray(byte** data, uint32_t& len)
 {
 	// it is thread safe this way.
 	uint32_t regionsize;
-	byte* regiondata = 0;
+	byte* regiondata = nullptr;
 	m_region.storeToByteArray(&regiondata, regionsize);
 
 	len = sizeof(id_type) + sizeof(uint32_t) + m_dataLength + regionsize;
@@ -372,10 +373,6 @@ SpatialIndex::RTree::RTree::RTree(IStorageManager& sm, Tools::PropertySet& ps) :
 	m_indexPool(100),
 	m_leafPool(100)
 {
-#ifdef HAVE_PTHREAD_H
-	pthread_mutex_init(&m_lock, NULL);
-#endif
-
 	Tools::Variant var = ps.getProperty("IndexIdentifier");
 	if (var.m_varType != Tools::VT_EMPTY)
 	{
@@ -397,10 +394,6 @@ SpatialIndex::RTree::RTree::RTree(IStorageManager& sm, Tools::PropertySet& ps) :
 
 SpatialIndex::RTree::RTree::~RTree()
 {
-#ifdef HAVE_PTHREAD_H
-	pthread_mutex_destroy(&m_lock);
-#endif
-
 	storeHeader();
 }
 
@@ -412,15 +405,11 @@ void SpatialIndex::RTree::RTree::insertData(uint32_t len, const byte* pData, con
 {
 	if (shape.getDimension() != m_dimension) throw Tools::IllegalArgumentException("insertData: Shape has the wrong number of dimensions.");
 
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	// convert the shape into a Region (R-Trees index regions only; i.e., approximations of the shapes).
 	RegionPtr mbr = m_regionPool.acquire();
 	shape.getMBR(*mbr);
 
-	byte* buffer = 0;
+	byte* buffer = nullptr;
 
 	if (len > 0)
 	{
@@ -436,10 +425,6 @@ bool SpatialIndex::RTree::RTree::deleteData(const IShape& shape, id_type id)
 {
 	if (shape.getDimension() != m_dimension) throw Tools::IllegalArgumentException("deleteData: Shape has the wrong number of dimensions.");
 
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	RegionPtr mbr = m_regionPool.acquire();
 	shape.getMBR(*mbr);
 	bool ret = deleteData_impl(*mbr, id);
@@ -448,13 +433,73 @@ bool SpatialIndex::RTree::RTree::deleteData(const IShape& shape, id_type id)
 }
 
 
-void SpatialIndex::RTree::RTree::containsWhatQuery(const IShape& query, IVisitor& v)
+void SpatialIndex::RTree::RTree::internalNodesQuery(const IShape& query, IVisitor& v)
 {
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("containsWhatQuery: Shape has the wrong number of dimensions.");
 
 #ifdef HAVE_PTHREAD_H
 	Tools::LockGuard lock(&m_lock);
 #endif
+
+	try
+	{
+		std::stack<NodePtr> st;
+		NodePtr root = readNode(m_rootID);
+		st.push(root);
+
+		while (! st.empty())
+		{
+			NodePtr n = st.top(); st.pop();
+
+			if(query.containsShape(n->m_nodeMBR))
+			{
+				IdVisitor vId = IdVisitor();
+				visitSubTree(n, vId);
+				const uint64_t nObj = vId.GetResultCount();
+				uint64_t *obj = new uint64_t[nObj];
+				std::copy(vId.GetResults().begin(), vId.GetResults().end(), obj);
+
+				Data data = Data(sizeof(uint64_t) * nObj, (byte *) obj, n->m_nodeMBR, n->getIdentifier());
+				v.visitData(data);
+				++(m_stats.m_u64QueryResults);
+			}
+			else
+			{
+				if(n->m_level == 0)
+				{
+					for (uint32_t cChild = 0; cChild < n->m_children; ++cChild)
+					{
+						if(query.containsShape(*(n->m_ptrMBR[cChild])))
+						{
+							Data data = Data(sizeof(id_type), (byte *) &n->m_pIdentifier[cChild], *(n->m_ptrMBR[cChild]), n->getIdentifier());
+							v.visitData(data);
+							++(m_stats.m_u64QueryResults);
+						}
+					}
+				}
+				else //not a leaf
+				{
+					if(query.intersectsShape(n->m_nodeMBR))
+					{
+						for (uint32_t cChild = 0; cChild < n->m_children; ++cChild)
+						{
+							st.push(readNode(n->m_pIdentifier[cChild]));
+						}
+					}
+				}
+			}
+		}
+
+	}
+	catch (...)
+	{
+		throw;
+	}
+}
+
+void SpatialIndex::RTree::RTree::containsWhatQuery(const IShape& query, IVisitor& v)
+{
+	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("containsWhatQuery: Shape has the wrong number of dimensions.");
 
 	try
 	{
@@ -521,13 +566,9 @@ void SpatialIndex::RTree::RTree::nearestNeighborQuery(uint32_t k, const IShape& 
 {
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("nearestNeighborQuery: Shape has the wrong number of dimensions.");
 
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	std::priority_queue<NNEntry*, std::vector<NNEntry*>, NNEntry::ascending> queue;
 
-	queue.push(new NNEntry(m_rootID, 0, 0.0));
+	queue.push(new NNEntry(m_rootID, nullptr, 0.0));
 
 	uint32_t count = 0;
 	double knearest = 0.0;
@@ -542,7 +583,7 @@ void SpatialIndex::RTree::RTree::nearestNeighborQuery(uint32_t k, const IShape& 
 
 		queue.pop();
 
-		if (pFirst->m_pEntry == 0)
+		if (pFirst->m_pEntry == nullptr)
 		{
 			// n is a leaf or an index.
 			NodePtr n = readNode(pFirst->m_id);
@@ -559,7 +600,7 @@ void SpatialIndex::RTree::RTree::nearestNeighborQuery(uint32_t k, const IShape& 
 				}
 				else
 				{
-					queue.push(new NNEntry(n->m_pIdentifier[cChild], 0, nnc.getMinimumDistance(query, *(n->m_ptrMBR[cChild]))));
+					queue.push(new NNEntry(n->m_pIdentifier[cChild], nullptr, nnc.getMinimumDistance(query, *(n->m_ptrMBR[cChild]))));
 				}
 			}
 		}
@@ -578,7 +619,7 @@ void SpatialIndex::RTree::RTree::nearestNeighborQuery(uint32_t k, const IShape& 
 	while (! queue.empty())
 	{
 		NNEntry* e = queue.top(); queue.pop();
-		if (e->m_pEntry != 0) delete e->m_pEntry;
+		if (e->m_pEntry != nullptr) delete e->m_pEntry;
 		delete e;
 	}
 }
@@ -596,10 +637,6 @@ void SpatialIndex::RTree::RTree::selfJoinQuery(const IShape& query, IVisitor& v)
 	if (query.getDimension() != m_dimension)
 		throw Tools::IllegalArgumentException("selfJoinQuery: Shape has the wrong number of dimensions.");
 
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	RegionPtr mbr = m_regionPool.acquire();
 	query.getMBR(*mbr);
 	selfJoinQuery(m_rootID, m_rootID, *mbr, v);
@@ -607,10 +644,6 @@ void SpatialIndex::RTree::RTree::selfJoinQuery(const IShape& query, IVisitor& v)
 
 void SpatialIndex::RTree::RTree::queryStrategy(IQueryStrategy& qs)
 {
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	id_type next = m_rootID;
 	bool hasNext = true;
 
@@ -689,6 +722,11 @@ void SpatialIndex::RTree::RTree::getIndexProperties(Tools::PropertySet& out) con
 	var.m_varType = Tools::VT_ULONG;
 	var.m_val.ulVal = m_pointPool.getCapacity();
 	out.setProperty("PointPoolCapacity", var);
+
+	var.m_varType = Tools::VT_LONGLONG;
+	var.m_val.llVal = m_headerID;
+	out.setProperty("IndexIdentifier", var);
+
 }
 
 void SpatialIndex::RTree::RTree::addCommand(ICommand* pCommand, CommandType ct)
@@ -696,13 +734,13 @@ void SpatialIndex::RTree::RTree::addCommand(ICommand* pCommand, CommandType ct)
 	switch (ct)
 	{
 		case CT_NODEREAD:
-			m_readNodeCommands.push_back(Tools::SmartPointer<ICommand>(pCommand));
+			m_readNodeCommands.push_back(std::shared_ptr<ICommand>(pCommand));
 			break;
 		case CT_NODEWRITE:
-			m_writeNodeCommands.push_back(Tools::SmartPointer<ICommand>(pCommand));
+			m_writeNodeCommands.push_back(std::shared_ptr<ICommand>(pCommand));
 			break;
 		case CT_NODEDELETE:
-			m_deleteNodeCommands.push_back(Tools::SmartPointer<ICommand>(pCommand));
+			m_deleteNodeCommands.push_back(std::shared_ptr<ICommand>(pCommand));
 			break;
 	}
 }
@@ -802,6 +840,11 @@ bool SpatialIndex::RTree::RTree::isIndexValid()
 void SpatialIndex::RTree::RTree::getStatistics(IStatistics** out) const
 {
 	*out = new Statistics(m_stats);
+}
+
+void SpatialIndex::RTree::RTree::flush()
+{
+	storeHeader();
 }
 
 void SpatialIndex::RTree::RTree::initNew(Tools::PropertySet& ps)
@@ -1144,7 +1187,7 @@ void SpatialIndex::RTree::RTree::storeHeader()
 void SpatialIndex::RTree::RTree::loadHeader()
 {
 	uint32_t headerSize;
-	byte* header = 0;
+	byte* header = nullptr;
 	m_pStorageManager->loadByteArray(m_headerID, headerSize, &header);
 
 	byte* ptr = header;
@@ -1194,7 +1237,7 @@ void SpatialIndex::RTree::RTree::insertData_impl(uint32_t dataLength, byte* pDat
 	assert(mbr.getDimension() == m_dimension);
 
 	std::stack<id_type> pathBuffer;
-	byte* overflowTable = 0;
+	byte* overflowTable = nullptr;
 
 	try
 	{
@@ -1252,10 +1295,10 @@ bool SpatialIndex::RTree::RTree::deleteData_impl(const Region& mbr, id_type id)
 		root.relinquish();
 	}
 
-	if (l.get() != 0)
+	if (l.get() != nullptr)
 	{
 		Leaf* pL = static_cast<Leaf*>(l.get());
-		pL->deleteData(id, pathBuffer);
+		pL->deleteData(mbr, id, pathBuffer);
 		--(m_stats.m_u64Data);
 		return true;
 	}
@@ -1340,7 +1383,7 @@ SpatialIndex::RTree::NodePtr SpatialIndex::RTree::RTree::readNode(id_type page)
 		else if (nodeType == PersistentLeaf) n = m_leafPool.acquire();
 		else throw Tools::IllegalStateException("readNode: failed reading the correct node type information");
 
-		if (n.get() == 0)
+		if (n.get() == nullptr)
 		{
 			if (nodeType == PersistentIndex) n = NodePtr(new Index(this, -1, 0), &m_indexPool);
 			else if (nodeType == PersistentLeaf) n = NodePtr(new Leaf(this, -1), &m_leafPool);
@@ -1390,10 +1433,6 @@ void SpatialIndex::RTree::RTree::deleteNode(Node* n)
 
 void SpatialIndex::RTree::RTree::rangeQuery(RangeQueryType type, const IShape& query, IVisitor& v)
 {
-#ifdef HAVE_PTHREAD_H
-	Tools::LockGuard lock(&m_lock);
-#endif
-
 	std::stack<NodePtr> st;
 	NodePtr root = readNode(m_rootID);
 
