@@ -52,8 +52,7 @@ Statistics::Statistics(const Statistics& s)
 }
 
 Statistics::~Statistics()
-{
-}
+= default;
 
 Statistics& Statistics::operator=(const Statistics& s)
 {
