@@ -35,12 +35,12 @@ namespace SpatialIndex
 		{
 		public:
 			DiskStorageManager(Tools::PropertySet&);
-			virtual ~DiskStorageManager();
+			~DiskStorageManager() override;
 
-			virtual void flush();
-			virtual void loadByteArray(const id_type page, uint32_t& len, byte** data);
-			virtual void storeByteArray(id_type& page, const uint32_t len, const byte* const data);
-			virtual void deleteByteArray(const id_type page);
+			void flush() override;
+			void loadByteArray(const id_type page, uint32_t& len, byte** data) override;
+			void storeByteArray(id_type& page, const uint32_t len, const byte* const data) override;
+			void deleteByteArray(const id_type page) override;
 
 		private:
 			class Entry
