@@ -38,12 +38,12 @@ namespace SpatialIndex
 		public:
 			MemoryStorageManager(Tools::PropertySet&);
 
-			virtual ~MemoryStorageManager();
+			~MemoryStorageManager() override;
 
-			virtual void flush();
-			virtual void loadByteArray(const id_type page, uint32_t& len, byte** data);
-			virtual void storeByteArray(id_type& page, const uint32_t len, const byte* const data);
-			virtual void deleteByteArray(const id_type page);
+			void flush() override;
+			void loadByteArray(const id_type page, uint32_t& len, byte** data) override;
+			void storeByteArray(id_type& page, const uint32_t len, const byte* const data) override;
+			void deleteByteArray(const id_type page) override;
 
 		private:
 			class Entry
@@ -52,7 +52,7 @@ namespace SpatialIndex
 				byte* m_pData;
 				uint32_t m_length;
 
-				Entry(uint32_t l, const byte* const d) : m_pData(0), m_length(l)
+				Entry(uint32_t l, const byte* const d) : m_pData(nullptr), m_length(l)
 				{
 					m_pData = new byte[m_length];
 					memcpy(m_pData, d, m_length);
